requirements.txt
openai>=1.12.0
python-docx>=1.1.0
pydantic>=2.6.0
python-dotenv>=1.0.1
streamlit>=1.31.0
watchdog>=4.0.0
google-genai>=0.1.0
mammoth>=1.6.0
